+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add `DMI_SIZE_MAX` constant for unknown `dmi_size_t` values
+
+### Fixed
+
+- Represent reserved minimum and maximum counts of chassis elements as unknown values
+- Show unknown firmware image size in firmware inventory information as unknown
+
 ## [0.4.0] - September 17, 2026
 
 ### Added